requirements.txt
numpy>=1.20
opencv-python>=4.5
Pillow>=10.0
pytesseract>=0.3.10